requirements.txt
discord.py
bs4
jinja2
ruamel.yaml
aiohttp
regex
pandas
seaborn